added smoothXYZDask
We are using this for smoothing the results
from fitLeicaScan.  We really don't want to smooth,
rather just regrid, but scipy.interpolate.interp2d
is not working for us, and this is.
With evenly spaced results from the ref and diff scan
we can then take the diff to find the bumps,
and just take zernike coefficients of this, since it's
evenly spaced.